feat: add enemy attacks, enemy types, enemy weapons (ZombieWeapons.h/.cpp); add player skills (Emperor's Divide, Call of The Forge God) (PlayerSkill.h/.cpp); add stunt effect (ZombieEffect.h/.cpp, PlayerEffect.h/.cpp); add skill cooldown bar (GameUI.cpp)
#TODO: add Boss

+++ sdl2training/SDL/src/GameFeatures/Global.cpp
@@ -14,7 +14,7 @@ float GetRandomFloat(float min_value, float max_value, float step)
 	return random_value;
 }
 
-int map(int x, int in_min, int in_max, int out_min, int out_max)
+float map(float x, float in_min, float in_max, float out_min, float out_max)
 {
 	return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
